% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.3M}
\alias{g.3M}
\title{An example DAG with mediators, where {I, M2, M3} are uninformative}
\format{
An object of class \code{dagitty} of length 1.
}
\usage{
g.3M
}
\description{
An example DAG with mediators, where {I, M2, M3} are uninformative
}
\keyword{datasets}
