% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.double}
\alias{g.double}
\title{A mirrored version of \code{g.example}. Variables {W4, W1} are uninformative.}
\format{
An object of class \code{dagitty} of length 1.
}
\usage{
g.double
}
\description{
A mirrored version of \code{g.example}. Variables {W4, W1} are uninformative.
}
\keyword{datasets}
