% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.example}
\alias{g.example}
\title{An example DAG, where {I, W4, W1} are uninformative}
\format{
An object of class \code{dagitty} of length 1.
}
\usage{
g.example
}
\description{
An example DAG, where {I, W4, W1} are uninformative
}
\keyword{datasets}
