% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.long}
\alias{g.long}
\title{A long example. Variables {I, W6, W2, W1} are uninformative.}
\format{
An object of class \code{dagitty} of length 1.
}
\usage{
g.long
}
\description{
A long example. Variables {I, W6, W2, W1} are uninformative.
}
\keyword{datasets}
