% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.strange.modified}
\alias{g.strange.modified}
\title{Modified version of \code{g.strange}. No variable is uninformative.}
\format{
An object of class \code{dagitty} of length 1.
}
\usage{
g.strange.modified
}
\description{
Modified version of \code{g.strange}. No variable is uninformative.
}
\keyword{datasets}
