% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.strange}
\alias{g.strange}
\title{A peculiar example. Variable M is uninformative.}
\format{
An object of class \code{dagitty} of length 1.
}
\usage{
g.strange
}
\description{
A peculiar example. Variable M is uninformative.
}
\keyword{datasets}
