% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.taxonomy}
\alias{g.taxonomy}
\title{Example of vertex taxonomy.}
\format{
An object of class \code{dagitty} of length 1.
}
\usage{
g.taxonomy
}
\description{
Example of vertex taxonomy.
}
\keyword{datasets}
