% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{getInformativeVariables}
\alias{getInformativeVariables}
\title{List all the informative variables of a DAG}
\usage{
getInformativeVariables(g)
}
\arguments{
\item{g}{A directed acyclic graph of class \code{dagitty}. A single vertex
should be labeled as \code{exposure} and another vertex should be labeled
as \code{outcome}.}
}
\value{
a character vector
}
\description{
This is the complement of \code{getUninformativeVariables}.
}
\seealso{
\code{\link{getUninformativeVariables}}
}
