% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{getUninformativeVariables}
\alias{getUninformativeVariables}
\title{List all the uninformative variables of a DAG}
\usage{
getUninformativeVariables(g)
}
\arguments{
\item{g}{A directed acyclic graph of class \code{dagitty}. A single vertex
should be labeled as \code{exposure} and another vertex should be labeled
as \code{outcome}.}
}
\value{
a character vector
}
\description{
This lists all the variables that are uninformative for estimating
the average effect of exposure on the outcome, given graph \code{g}.
}
\examples{
data("g.example")
getUninformativeVariables(g.example)
}
\seealso{
\code{\link[dagitty]{dagitty}}
}
