% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformula.R
\name{gFormula}
\alias{gFormula}
\title{The g-formula for identifying the counterfactual mean}
\usage{
gFormula(g)
}
\arguments{
\item{g}{A directed acyclic graph of class \code{dagitty}. A single vertex
should be labeled as \code{exposure} and another vertex should be labeled
as \code{outcome}. Use \code{\link[dagitty]{dagitty}} to create this object.}
}
\value{
Character string for the formula
}
\description{
The g-formula (Robins, 1986) for the counterfactual mean of the outcome when
level \code{a} is imposed on the exposure.
}
\examples{
data("g.example")
# compare:
gFormula(g.example)
gFormula(reduceDAG(g.example, verbose=FALSE))
}
\references{
James M. Robins. "A new approach to causal inference in mortality studies with
a sustained exposure period—application to control of the healthy worker
survivor effect." Mathematical modelling 7.9-12 (1986): 1393-1512.

Thomas S. Richardson and James M. Robins. "Single world intervention graphs
(SWIGs): A unification of the counterfactual and graphical approaches to
causality." Center for the Statistics and the Social Sciences, University
of Washington Series. Working Paper 128.30 (2013): 2013.
}
