% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R, R/reduceDAG.R
\docType{package}
\name{reduceDAG}
\alias{reduceDAG}
\title{Reduce a causal directed acyclic graph (DAG)}
\usage{
reduceDAG(g, verbose = TRUE)
}
\arguments{
\item{g}{A directed acyclic graph of class \code{dagitty}. A single vertex
should be labeled as \code{exposure} and another vertex should be labeled
as \code{outcome}. Use \code{\link[dagitty]{dagitty}} to create this object.}

\item{verbose}{If \code{TRUE}, print the variables eliminated and the
reduced g-formula, also plot the reduced graph. (default: \code{TRUE})}
}
\value{
a DAG of class \code{daggity}
}
\description{
Reduce the causal DAG to a (potentially) smaller DAG, of which all the
variables are informative for estimating the effect of exposure on the
outcome.

Consider estimating a counterfactual mean of a point intervention
(or the associated average treatment effect) when the causal structure is known and
described by a directed acyclic graph (DAG). Suppose there is no unobserved
confounder. It turns out that not every variable in the graph is informative for
optimally estimating the counterfactual mean. This package identifies and
projects out all the uninformative variables from the DAG. Based on the reduced
DAG, an efficient estimator achieves the semiparametric efficiency bound with
respect to the original DAG. In particular, the g-formula (Robins, 1986)
prescribed by the reduced formula, when estimated with empirical plugin, is
efficient.
}
\details{
Use \code{\link[dagitty]{dagitty}} to represent a DAG, of which the
\code{exposure} vertex and the \code{outcome} vertex must be labeled.

Use \code{\link{reduceDAG}} to reduce a DAG.

Use \code{\link{gFormula}} to display the g-formula associated with a DAG.

Use \code{\link{getUninformativeVariables}} to list all the uninformative
variables that can be eliminated from a DAG.
}
\examples{
data("g.example")
reduceDAG(g.example)
reduceDAG(g.example, verbose=FALSE)
data("g.long")
g <- reduceDAG(g.long, verbose=FALSE)
cat(gFormula(g))
plot(g)
}
\seealso{
\code{\link{gFormula}}, \code{\link{getUninformativeVariables}},
\code{\link{getInformativeVariables}}

\code{\link[dagitty]{dagitty}}, \code{\link[dagitty]{exposures}},
\code{\link[dagitty]{outcomes}}
}
